Add class for informations, separate class for each flowers
I needed the pair of text and img for each flower to be bundled together so that the text appears below the img when I apply "display: flex;" to the parent class (written below this paragraph). So I add a different class for each flower:
.barberry for barberry
.chilli for chilli
.pepper for pepper
.saffron for saffron

I then grouped the 4 classes together into a class called "plants-information" so that they could be displayed side by side when I apply "display: flex;" to .plants-information.